chore: configure build for all ui packages
