+++ packages/ui/jest.config.js
@@ -5,24 +5,23 @@ module.exports = {
     '<rootDir>/libs/ui/utils',
     '<rootDir>/libs/ui/localization',
     '<rootDir>/libs/ui/theme',
-    '<rootDir>/libs/ui/state-management',
+    '<rootDir>/libs/ui/shared-state',
     '<rootDir>/libs/ui/assets',
-    '<rootDir>/libs/certificate/data-access',
-    '<rootDir>/libs/exchange/data-access',
+    '<rootDir>/libs/ui/blockchain',
+    '<rootDir>/libs/exchange/data',
     '<rootDir>/libs/exchange/view',
     '<rootDir>/libs/exchange/logic',
+    '<rootDir>/libs/certificate/data',
     '<rootDir>/libs/certificate/view',
     '<rootDir>/libs/certificate/logic',
-    '<rootDir>/libs/user/data-access',
+    '<rootDir>/libs/user/data',
     '<rootDir>/libs/user/view',
     '<rootDir>/libs/user/logic',
+    '<rootDir>/libs/device/data',
     '<rootDir>/libs/device/logic',
     '<rootDir>/libs/device/view',
     '<rootDir>/libs/organization/view',
     '<rootDir>/libs/organization/data',
     '<rootDir>/libs/organization/logic',
-    '<rootDir>/libs/ui/api-clients',
-    '<rootDir>/libs/device/data',
-    '<rootDir>/libs/ui/blockchain',
   ],
 };
