+++ packages/ui/libs/certificate/data/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@energyweb/origin-ui-certificate-data-access",
+  "name": "@energyweb/origin-ui-certificate-data",
   "version": "0.0.1",
   "repository": {
     "type": "git",
