+++ packages/ui/libs/device/logic/src/registerForm/index.ts
@@ -1,1 +1,2 @@
 export * from './form';
+export * from './types';
