+++ packages/ui/libs/device/view/src/containers/smartMeter/index.ts
@@ -1,0 +1,1 @@
+export * from './SmartMeterBlock';
