+++ packages/ui/libs/exchange/data/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@energyweb/origin-ui-exchange-data-access",
+  "name": "@energyweb/origin-ui-exchange-data",
   "version": "0.0.1",
   "repository": {
     "type": "git",
