+++ packages/ui/libs/ui/core/src/components/index.ts
@@ -8,3 +8,4 @@ export * from './text';
 export * from './notification';
 export * from './list';
 export * from './card';
+export * from './buttons';
