--- packages/ui/libs/ui/utils/src/date/formatDate.spec.ts
@@ -1,7 +1,0 @@
-import { formatDate } from './formatDate';
-
-describe('time', function () {
-  it('should format date properly', function () {
-    console.log(formatDate(Date.now(), true));
-  });
-});
