+++ packages/ui/nx.json
@@ -31,22 +31,22 @@
     "ui-localization": { "tags": [] },
     "ui-theme": { "tags": [] },
     "ui-assets": { "tags": [] },
+    "ui-blockchain": { "tags": [] },
+    "ui-shared-state": { "tags": [] },
+    "exchange-data": { "tags": [] },
     "exchange-view": { "tags": [] },
     "exchange-logic": { "tags": [] },
     "certificate-view": { "tags": [] },
     "certificate-logic": { "tags": [] },
+    "certificate-data": { "tags": [] },
+    "user-data": { "tags": [] },
     "user-view": { "tags": [] },
     "user-logic": { "tags": [] },
+    "device-data": { "tags": [] },
     "device-logic": { "tags": [] },
     "device-view": { "tags": [] },
-    "ui-shared-state": { "tags": [] },
     "organization-view": { "tags": [] },
     "organization-data": { "tags": [] },
-    "organization-logic": { "tags": [] },
-    "device-data": { "tags": [] },
-    "certificate-data": { "tags": [] },
-    "ui-blockchain": { "tags": [] },
-    "user-data": { "tags": [] },
-    "exchange-data": { "tags": [] }
+    "organization-logic": { "tags": [] }
   }
 }
