+++ packages/ui/package.json
@@ -5,8 +5,21 @@
   "scripts": {
     "nx": "nx",
     "start": "nx serve",
-    "build": "yarn config-env && nx run ui-utils:build && nx run ui-theme:build && nx run ui-assets:build && nx run ui-localization:build && nx run ui-shared-state:build && nx run ui-core:build && nx run ui-blockchain:build && nx build",
     "config-env": "bash env.sh",
+    "build:ui-utils": "nx run ui-utils:build --buildableProjectDepsInPackageJsonType=dependencies",
+    "build:ui-theme": "nx run ui-theme:build",
+    "build:ui-assets": "nx run ui-assets:build",
+    "build:ui-localization": "nx run ui-localization:build",
+    "build:ui-shared-state": "nx run ui-shared-state:build --buildableProjectDepsInPackageJsonType=dependencies",
+    "build:ui-core": "nx run ui-core:build --buildableProjectDepsInPackageJsonType=dependencies",
+    "build:ui-blockchain": "nx run ui-blockchain:build",
+    "build:user-ui": "nx run user-data:build && nx run user-logic:build && nx run user-view:build",
+    "build:organization-ui": "nx run organization-data:build && nx run organization-logic:build && nx run organization-view:build",
+    "build:certificate-ui": "nx run certificate-data:build && nx run certificate-logic:build && nx run certificate-view:build",
+    "build:device-ui": "nx run device-data:build && nx run device-logic:build && nx run device-view:build",
+    "build:exchange-ui": "nx run exchange-data:build && nx run exchange-logic:build && nx run exchange-view:build",
+    "build:all-packages": "yarn build:ui-utils && yarn build:ui-theme && yarn build:ui-assets && yarn build:ui-localization && yarn build:ui-shared-state && yarn build:ui-core && yarn build:ui-blockchain && yarn build:user-ui && yarn build:organization-ui && yarn build:certificate-ui && yarn build:device-ui && yarn build:exchange-ui",
+    "build": "yarn config-env && yarn build:all-packages && nx build",
     "test": "nx test",
     "lint": "nx workspace-lint && nx lint",
     "e2e": "nx e2e",
@@ -41,35 +54,35 @@
     "@energyweb/origin-backend-core": "8.0.1",
     "@energyweb/utils-general": "11.0.2",
     "@energyweb/issuer": "3.2.0",
-    "@hookform/resolvers": "^2.3.0",
+    "@hookform/resolvers": "2.3.0",
     "@material-ui/core": "5.0.0-alpha.29",
     "@material-ui/icons": "5.0.0-alpha.29",
     "@material-ui/lab": "5.0.0-alpha.32",
     "@material-ui/styles": "5.0.0-alpha.29",
     "@material-ui/system": "5.0.0-alpha.29",
-    "@react-google-maps/api": "^2.1.1",
+    "@react-google-maps/api": "2.1.1",
     "axios": "0.21.1",
     "core-js": "3.6.5",
-    "dayjs": "^1.10.4",
+    "dayjs": "1.10.4",
     "document-register-element": "1.13.1",
     "ethers": "5.3.1",
     "history": "5.0.0",
-    "i18next": "^20.2.1",
+    "i18next": "20.2.1",
     "i18next-icu": "1.4.2",
     "lodash": "4.17.21",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-dropzone": "^11.3.2",
-    "react-error-boundary": "^3.1.1",
-    "react-hook-form": "^7.1.0",
-    "react-i18next": "^11.8.13",
+    "react-dropzone": "11.3.2",
+    "react-error-boundary": "3.1.1",
+    "react-hook-form": "7.1.0",
+    "react-i18next": "11.8.13",
     "react-is": "17.0.2",
-    "react-material-ui-carousel": "^2.2.4",
+    "react-material-ui-carousel": "2.2.4",
     "react-query": "3.15.1",
     "react-router": "6.0.0-beta.0",
     "react-router-dom": "6.0.0-beta.0",
     "react-toastify": "7.0.4",
-    "react-use": "^17.2.4",
+    "react-use": "17.2.4",
     "regenerator-runtime": "0.13.7",
     "react-chartjs-2": "3.0.3",
     "@web3-react/core": "6.1.9",
@@ -77,11 +90,11 @@
     "@ethersproject/providers": "5.3.1",
     "chart.js": "3.3.2",
     "tslib": "2.3.0",
-    "yup": "^0.32.9"
+    "yup": "0.32.9"
   },
   "devDependencies": {
     "@babel/core": "7.9.6",
-    "@babel/preset-env": "^7.13.15",
+    "@babel/preset-env": "7.13.15",
     "@babel/preset-react": "7.9.4",
     "@babel/preset-typescript": "7.9.0",
     "@emotion/babel-preset-css-prop": "11.0.0",
@@ -95,48 +108,48 @@
     "@nrwl/tao": "12.6.5",
     "@nrwl/web": "12.6.5",
     "@nrwl/workspace": "12.6.5",
-    "@storybook/addon-essentials": "^6.2.9",
-    "@storybook/addon-knobs": "^6.2.9",
-    "@storybook/react": "^6.2.9",
+    "@storybook/addon-essentials": "6.2.9",
+    "@storybook/addon-knobs": "6.2.9",
+    "@storybook/react": "6.2.9",
     "@svgr/webpack": "5.5.0",
     "@svgr/rollup": "5.5.0",
-    "@testing-library/jest-dom": "^5.11.10",
+    "@testing-library/jest-dom": "5.11.10",
     "@testing-library/react": "11.1.2",
     "@types/jest": "26.0.24",
-    "@types/lodash": "^4.14.168",
+    "@types/lodash": "4.14.168",
     "@types/material-ui": "^0.21.8",
     "@types/node": "14.17.7",
     "@types/react": "17.0.1",
     "@types/react-dom": "17.0.1",
     "@types/react-is": "17.0.0",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react-router-dom": "5.1.7",
     "@types/webpack": "4.41.21",
     "@typescript-eslint/eslint-plugin": "4.3.0",
     "@typescript-eslint/parser": "4.3.0",
     "babel-jest": "26.2.2",
     "babel-loader": "8.1.0",
-    "babel-plugin-styled-components": "^1.12.0",
-    "cypress": "^7.3.0",
+    "babel-plugin-styled-components": "1.12.0",
+    "cypress": "7.3.0",
     "eslint": "7.10.0",
     "eslint-config-airbnb": "18.2.1",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "eslint-plugin-cypress": "2.10.3",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-prettier": "3.3.1",
     "eslint-plugin-react": "7.21.5",
     "eslint-plugin-react-hooks": "4.2.0",
     "jest": "27.0.6",
-    "jest-fetch-mock": "^3.0.3",
-    "jest-localstorage-mock": "^2.4.9",
+    "jest-fetch-mock": "3.0.3",
+    "jest-localstorage-mock": "2.4.9",
     "prettier": "2.3.2",
-    "test-react-hooks": "^2.0.4",
+    "test-react-hooks": "2.0.4",
     "ts-jest": "26.4.0",
     "ts-node": "9.1.1",
     "typescript": "4.3.5",
-    "url-loader": "^3.0.0",
+    "url-loader": "3.0.0",
     "webpack": "4.42.0",
-    "webpack-bundle-analyzer": "^4.4.1",
+    "webpack-bundle-analyzer": "4.4.1",
     "tsconfig-paths-webpack-plugin": "3.5.1"
   },
   "workspaces": [
